Move renameSlot and renameAttrib to NodeProc
