+++ packages/core/src/html/NodeProcessor.js
@@ -23,6 +23,7 @@ const { ComponentProcessor } = require('./ComponentProcessor');
 const { shiftSlotNodeDeeper, transformOldSlotSyntax } = require('./vueSlotSyntaxProcessor');
 const { warnBodyTag, warnConflictingAtributesMap, warnDeprecatedAtributesMap } = require('./warnings');
 const { processScriptTag, processStyleTag } = require('./scriptAndStyleTagProcessor');
+const { getVslotShorthandName } = require('./vueSlotSyntaxProcessor');
 
 const utils = require('../utils');
 const logger = require('../utils/logger');
@@ -79,6 +80,26 @@ class NodeProcessor {
 
   static _isText(node) {
     return node.type === 'text' || node.type === 'comment';
+  }
+
+  static _renameSlot(node, originalName, newName) {
+    if (node.children) {
+      node.children.forEach((child) => {
+        const vslotShorthandName = getVslotShorthandName(child);
+        if (vslotShorthandName && vslotShorthandName === originalName) {
+          const newVslot = `#${newName}`;
+          child.attribs[newVslot] = '';
+          delete child.attribs[`#${vslotShorthandName}`];
+        }
+      });
+    }
+  }
+
+  static _renameAttribute(node, originalAttribute, newAttribute) {
+    if (_.has(node.attribs, originalAttribute)) {
+      node.attribs[newAttribute] = node.attribs[originalAttribute];
+      delete node.attribs[originalAttribute];
+    }
   }
 
   /*
@@ -166,7 +187,12 @@ class NodeProcessor {
         break;
       case 'modal':
         this.componentProcessor.processModalAttributes(node);
-        BootstrapVueProcessor.processModalAttributes(node);
+        BootstrapVueProcessor.processModalAttributes(node,
+                                                     (originalName, newName) =>
+                                                       NodeProcessor._renameSlot(node, originalName, newName),
+                                                     (originalAttribute, newAttribute) =>
+                                                       // eslint-disable-next-line max-len
+                                                       NodeProcessor._renameAttribute(node, originalAttribute, newAttribute));
         break;
       case 'tab':
       case 'tab-group':
